reduce line len

+++ pdc_dp_means/dpmeans.py
@@ -82,7 +82,9 @@ def _dpmeans_single_lloyd(
             if max_clusters is None or max_clusters > centers.shape[0]:
                 if max_index[0] != -1 and max_distance[0] > delta:
                     centers = np.vstack((centers, X[max_index])).astype(X.dtype)
-                    centers_new = np.vstack((centers_new, X[max_index])).astype(X.dtype)
+                    centers_new = np.vstack((centers_new, X[max_index])).astype(
+                        X.dtype
+                    )
                     weight_in_clusters = np.hstack([weight_in_clusters, [0]]).astype(
                         X.dtype
                     )
@@ -1009,7 +1011,9 @@ class MiniBatchDPMeans(KMeans):
             # Perform the iterative optimization until convergence
             for i in range(n_steps):
                 # Sample a minibatch from the full dataset
-                minibatch_indices = random_state.randint(0, n_samples, self._batch_size)
+                minibatch_indices = random_state.randint(
+                    0, n_samples, self._batch_size
+                )
                 tic = time()
                 # Perform the actual update step on the minibatch data
                 (
